move attachment_preview_pdf.ts to extension/js/common/ui/
